move text parsing utilities to 'Misc Functions'
A while ago some text file parsing functions were introduced in the TPS
pencil beam actor. These utilities were recently copied into the dose
actor. When a utility is useful in multiple source files then it's
better to store it in a more central place, such as
general/*/GateMiscfunctions.*, where similar utilities are also kept.

The utilities in question do the following:
- read the "next content line", skipping empty lines and comment lines
- read a content line as a vector of N items of type T, for any number N
  and type T. Raise an error at an unexpected number of items or when
  the input text cannot be streamed into an object of the desired type.
