+++ source/general/include/GateMiscFunctions.hh
@@ -19,6 +19,9 @@ See GATE/LICENSE.txt for further details
 #include <iostream>
 #include <fstream>
 #include <string>
+#include <sstream>
+#include <iterator>
+#include <exception>
 
 #include "G4UIcommand.hh"
 #include "G4VSolid.hh"
@@ -150,6 +153,27 @@ int GetIndexFromTime(std::vector<double> & mTimeList, double aTime);
 
 //-----------------------------------------------------------------------------
 G4String GetSaveCurrentFilename(G4String & mSaveFilename);
+
+//------------------------------------------------------------------------------------------------------
+//  try get N values of type T from a given input line
+// * throw exception with informative error message in case of trouble.
+// * NOTE that while this catches some common errors, it is not yet fool proof.
+template<typename T, int N>
+typename std::vector<T> parse_N_values_of_type_T(std::string line,int lineno, const std::string& fname);
+
+//------------------------------------------------------------------------------------------------------
+// Function to read the next content line
+// * skip all comment lines (lines string with a '#')
+// * skip empty
+// * throw exception with informative error message in case of missing data
+std::string ReadNextContentLine( std::istream& input, int& lineno, const std::string& fname );
+
+//------------------------------------------------------------------------------------------------------
+// Function to read AND parse the next content line
+// * check that we really get N values of type T from the current line
+template<typename T, int N>
+typename std::vector<T>  ParseNextContentLine( std::istream& input, int& lineno, const std::string& fname );
+
 
 
 #include "GateMiscFunctions.icc"
